feat: Make API call to github servers
In this a commit, an API GET request was made to the github servers using axios.
the API call is made from within the fetchGithubInformation() function. within
the then() promise method, the created dom components are then populated with
the response data from the axios get request.

Note that, at this point, the components being created is now done within the
then() promise call back function.