feat: Add DOM cards for list of followers Array
In this commit, I added cards for each individual present in the followers array provided.
to do this, I manually added followers to the followersArray and looped through each item.
for each follower in the array, I made an API call using axios and called the fetchGithub-
Information() function which then gets the follower's info and adds a card to the DOM for
that particular follower.

the axios get request was also extracted from the fetchGithubInformation() method in just
to allow for its reusability.